Add unit testing

+++ src/main/java/com/team2502/ezauton/utils/MathUtils.java
@@ -534,7 +534,7 @@ public final class MathUtils
             return immutableVector;
         }
 
-        public static ImmutableVector absoluteToRelativeCoord(ImmutableVector coordinateAbsolute, ImmutableVector robotCoordAbs, float robotHeading)
+        public static ImmutableVector absoluteToRelativeCoord(ImmutableVector coordinateAbsolute, ImmutableVector robotCoordAbs, double robotHeading)
         { return rotate2D(coordinateAbsolute.sub(robotCoordAbs), -robotHeading); }
     }
